refactor: apis 문서 정리
axios api 코드 정리

#69
