+++ FE/src/apis/apis.ts
@@ -1,28 +1,10 @@
-import { HomeFeedProps } from "../hooks/useFeeds";
-//TODO: 반환 타입 지정하기
-
-import {
-  fundingDetailURL,
-  homeURL,
-  joinURL,
-  loginURL,
-  univURL,
-} from "@/constants/urls";
+import { joinURL, loginURL, headerMemberUrl } from "@/constants/urls";
 import { Server } from "./axios";
 import { TOKEN_EXPIRED_TIME } from "@/constants/LoginData";
-import { HomeFeed, UserFeed } from "@/Model/Feed";
-import { Univs } from "@/Model/Univ";
-import { TPostResponse, TResponse } from "@/Model/Response";
-import { UnivState } from "@/store/HeaderState";
-import {
-  Funding,
-  FundingDetail,
-  FundingOptions,
-  Pledge,
-} from "@/Model/Funding";
+import { UserFeed } from "@/Model/Feed";
 import { LoginUser, UserDetail, UserForm } from "@/Model/User";
 import { Backed } from "@/Model/Backed";
-import { Coin } from "@/Model/Coin";
+import { TResponse } from "@/Model/Response";
 
 export const checkDuplicate = async (email: string) => {
   const data = await Server.get(joinURL, { params: { email } }).then(
@@ -75,87 +57,13 @@ export const login = async (submitData: LoginUser) => {
     });
 };
 
-export const getHomeFeed = async ({
-  category,
-  sort,
-  uni,
-  q,
-}: HomeFeedProps): Promise<TResponse<HomeFeed>> => {
-  const data = await Server.get(homeURL, {
-    params: { category, sort, uni, q },
-  }).then((res) => res.data);
-  return data;
-};
-
-export const getUnivList = async ({
-  region,
-  q,
-}: UnivState): Promise<TResponse<Univs>> => {
-  console.log(region);
-  const data = Server.get(univURL, {
-    params: {
-      region,
-      q,
-    },
-  }).then((res) => res.data);
-  return data;
-};
-
-export const getFundingDetail = async (
-  fundingId: number
-): Promise<TResponse<FundingDetail>> => {
-  //TODO: 반환 타입 지정하기
-  const data = Server.get(fundingDetailURL + fundingId).then((res) => res.data);
-  return data;
-};
-
-export const getFundingLongDescription = async (
-  fundingId: number
-): Promise<TResponse<{ longDescription: string }>> => {
-  const data = Server.get(fundingDetailURL + fundingId + "/description").then(
-    (res) => res.data
-  );
-  return data;
-};
-
-//TODO: url 전역변수로 관리하기!!!!!
-
 export const getUser = async (): Promise<
   TResponse<{
     memberId: number;
     image: string;
   }>
 > => {
-  const data = Server.get("/member").then((res) => res.data);
-  return data;
-};
-
-export const getPledge = async (
-  fundingId: number
-): Promise<TResponse<Pledge>> => {
-  const pledgeUrl = `/funding/${fundingId}/payment`;
-  const data = Server.get(pledgeUrl).then((res) => res.data);
-  return data;
-};
-
-export const getCharge = async (): Promise<TResponse<Coin>> => {
-  const data = Server.get("/charge").then((res) => res.data);
-  return data;
-};
-
-export const postPayment = async (
-  memberId: number,
-  type: "DONATE" | "GIFT",
-  option: FundingOptions[],
-  price: number
-): Promise<TPostResponse> => {
-  const body = {
-    memberId,
-    type,
-    option,
-    price,
-  };
-  const data = Server.post(`/funding/payment`, body).then((res) => res.data);
+  const data = Server.get(headerMemberUrl).then((res) => res.data);
   return data;
 };
 
@@ -182,13 +90,5 @@ export const getBacked = async (
   const data = await Server.get(`/member/${memberId}/backed/${fundingId}`).then(
     (res) => res.data
   );
-  return data;
-};
-
-export const postCharge = async (coin: number): Promise<TPostResponse> => {
-  const data = await Server.post("/charge", {
-    coin,
-  }).then((res) => res.data);
-  console.log(data);
   return data;
 };
