--- FE/src/app/api/funding/payment/route.ts
@@ -1,11 +1,0 @@
-export async function POST(request: Request) {
-  const body = await request.json();
-  console.log(body);
-  return new Response(
-    JSON.stringify({
-      success: true,
-      response: null,
-      error: null,
-    })
-  );
-}
