+++ FE/src/components/fundingDetail/Info/FundingDetailBtnSection.tsx
@@ -14,8 +14,8 @@ export default function FundingDetailBtnSection({
   const option = useRecoilValue(SelectedOptions);
   const isOkToPledge = option.length > 0;
   return (
-    <section className="flex gap-[0.63rem]">
-      <button className="w-[4.375rem] h-[4.375rem] shrink-0 rounded-[0.25rem] flex justify-center items-center border mt-4 ">
+    <section className="flex gap-[0.63rem] mt-4">
+      <button className="w-[4.375rem] h-[4.375rem] shrink-0 rounded-[0.25rem] flex justify-center items-center border ">
         <Share />
       </button>
       <StyledBtn
