+++ FE/src/components/pledge/CostInfo.tsx
@@ -2,13 +2,16 @@ import Rectangle from "../common/icons/Rectangle";
 import Payment from "./Payment";
 import PaymentBg from "./PaymentBg";
 
-export default function CostInfo({
-  type,
-}: Readonly<{ type: "DONATE" | "GIFT" }>) {
+type Props = {
+  fundingId: number;
+  type: "DONATE" | "GIFT";
+};
+
+export default function CostInfo({ fundingId, type }: Readonly<Props>) {
   return (
     <section className="relative px-5 ">
       <Rectangle />
-      <Payment type={type} />
+      <Payment type={type} fundingId={fundingId} />
       <PaymentBg />
     </section>
   );
