+++ FE/src/constants/urls.ts
@@ -3,3 +3,5 @@ export const loginURL = "/member/login";
 export const homeURL = "home";
 export const univURL = "/uni";
 export const fundingDetailURL = "/funding/";
+export const headerMemberUrl = "/member";
+export const userCoinUrl = "/charge";
